Fix and various changes to all benchmarks
It seems that we are not able to let one process allocate the
cohort lock (argo::new_*), or even this allocation being done
collectively (argo::conew_*), due to the global flag that is
allocated collectively internally. Besides changing the cohort
lock allocations, some data structures that were not necessary
to be allocated globally, they've been brough to local memory,
and also team process initialization was added whenever possible.